+++ foxy_gh_farmer/gigahorse_binary_manager.py
@@ -8,9 +8,11 @@ from sys import platform
 from tempfile import TemporaryDirectory
 from zipfile import ZipFile
 
-from aiohttp import ClientSession
+from aiohttp import ClientSession, ClientTimeout
 from chia.server.server import ssl_context_for_root
 from chia.ssl.create_ssl import get_mozilla_ca_crt
+from yaspin import yaspin
+from yaspin.core import Yaspin
 
 _gigahorse_release = "2.1.3.giga26"
 _gigahorse_release_archive_base = f"chia-gigahorse-farmer-{_gigahorse_release}"
@@ -29,13 +31,15 @@ class GigahorseBinaryManager:
         if gigahorse_path.exists():
             return gigahorse_path
 
-        self._logger.info(f"Downloading Gigahorse {_gigahorse_release} ..")
-        self._cache_path.mkdir(parents=True, exist_ok=True)
-        archive_file_name = self._get_archive_file_name()
-        with TemporaryDirectory() as temp_dir:
-            archive_path = join(temp_dir, archive_file_name)
-            await self._download_release(archive_path)
-            self._extract_file(archive_path, gigahorse_base_path)
+        with yaspin(text=f"Preparing to download Gigahorse {_gigahorse_release} ..") as spinner:
+            self._cache_path.mkdir(parents=True, exist_ok=True)
+            archive_file_name = self._get_archive_file_name()
+            with TemporaryDirectory() as temp_dir:
+                archive_path = join(temp_dir, archive_file_name)
+                await self._download_release(archive_path, spinner=spinner)
+                spinner.text = f"Extracting Gigahorse {_gigahorse_release} .."
+                self._extract_file(archive_path, gigahorse_base_path)
+        self._logger.info(f"✅ Downloaded Gigahorse {_gigahorse_release}")
 
         return gigahorse_path
 
@@ -54,13 +58,20 @@ class GigahorseBinaryManager:
 
         raise RuntimeError(f"Can not extract {archive_file_path}, unsupported extension")
 
-    async def _download_release(self, to_path: str):
-        chunk_size = 5 * 2**20  # MB
-        async with ClientSession() as client:
+    async def _download_release(self, to_path: str, spinner: Yaspin):
+        one_mib_in_bytes = 2 ** 20
+        chunk_size = 5 * one_mib_in_bytes
+        downloaded_size_mib = 0
+        async with ClientSession(timeout=ClientTimeout(total=15 * 60, connect=60)) as client:
             async with client.get(self._get_release_download_url(), ssl=self._ssl_context) as res:
+                total_bytes = int(res.headers.get('content-length'))
+                total_mib = total_bytes / one_mib_in_bytes
                 with open(to_path, 'wb') as fd:
                     async for chunk in res.content.iter_chunked(chunk_size):
                         fd.write(chunk)
+                        downloaded_size_mib += len(chunk) / one_mib_in_bytes
+                        percentage = (downloaded_size_mib / total_mib) * 100
+                        spinner.text = f"Downloading Gigahorse {_gigahorse_release} ({downloaded_size_mib:.2f}/{total_mib:.2f} MiB, {percentage:.2f}%) .."
 
     def _get_release_download_url(self):
         return f"{_download_url_base}/{_gigahorse_release}/{self._get_archive_file_name()}"